move wrapper stuff to new cache package
